+++ pages/home.js
@@ -1,7 +1,8 @@
 import { useRouter } from "next/router";
 import Head from "next/head";
-import en from "../locales/en/lodging.json";
-import pt from "../locales/pt/lodging.json";
+
+import en from "../locales/en/pages/lodging.json";
+import pt from "../locales/pt/pages/lodging.json";
 
 import SocialButton from "../components/atoms/SocialButton";
 
@@ -45,6 +46,7 @@ const Home = () => {
           />
           <SocialButton
             theme="instagram"
+            s
             label="Instagram"
             href="https://instagram.com/quintadonairia"
             icon="instagram"
