+++ pages/index.js
@@ -1,15 +1,14 @@
 import { useRouter } from "next/router";
-import en from "../locales/en/index.json";
-import pt from "../locales/pt/index.json";
-
 import Head from "next/head";
 import Link from "next/link";
 
+import en from "../locales/en/pages/index.json";
+import pt from "../locales/pt/pages/index.json";
+
+import Button from "../components/atoms/Button";
 import Banner from "../components/molecules/Banner";
 import Header from "../components/molecules/Header";
 import Gallery from "../components/organisms/Gallery";
-
-import Button from "../components/atoms/Button";
 import Hero from "../components/organisms/Hero";
 
 const Home = () => {
